Extract CLI commands into separate files
