+++ cli_createblockchain.go
@@ -1,0 +1,9 @@
+package main
+
+import "fmt"
+
+func (cli *CLI) createBlockchain(address string) {
+	bc := CreateBlockchain(address)
+	bc.db.Close()
+	fmt.Println("Done!")
+}
