+++ cli_createwallet.go
@@ -1,0 +1,11 @@
+package main
+
+import "fmt"
+
+func (cli *CLI) createWallet() {
+	wallets, _ := NewWallets()
+	address := wallets.CreateWallet()
+	wallets.SaveToFile()
+
+	fmt.Printf("Your new address: %s\n", address)
+}
