+++ cli_listaddress.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"fmt"
+	"log"
+)
+
+func (cli *CLI) listAddresses() {
+	wallets, err := NewWallets()
+	if err != nil {
+		log.Panic(err)
+	}
+	addresses := wallets.GetAddresses()
+
+	for _, address := range addresses {
+		fmt.Println(address)
+	}
+}
